:art: move handlers to their respective folder
